third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-density-and-width-descriptors-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("foo").clientWidth==100 is true
This test passes if the image below is a blue square, since the srcset resource doesn't load, as it has both 'w' and 'x' descriptors.
